requirements.txt
argcomplete==1.12.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
colorama==0.4.5
commitizen==2.29.2
commitizen-emoji==0.1.5
cryptography==37.0.4
decli==0.5.2
defusedxml==0.7.1
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==13.15.1
gunicorn==20.1.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
questionary==1.10.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tomlkit==0.11.1
toposort==1.7
typing-extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
wcwidth==0.2.5
whitenoise==6.2.0
